tornado server for django app, media/static files

+++ Notes/views.py
@@ -62,7 +62,10 @@ def notes(request, id=None):
             note = Message.objects.filter(id=id)[0]
         except IndexError:
             return redirect("index")
-        if request.method == "DELETE":
+        if request.method == "POST" and request.body == b"DELETE":
+            '''
+                USING POST REQUEST FOR DELETION BECAUSE TORNADO SERVER FAILING TO HANDLE DELETE REQUEST
+            '''
             if note.image:
                 note.image.delete()
             note.delete()
